Fixed payment is working

+++ Repository/Models/Customer.cs
@@ -4,9 +4,9 @@
     {
         public string Name { get; set; }
 
-        public decimal TotalLoan { get; set; }
+        public double TotalLoan { get; set; }
 
-        public decimal Interest { get; set; }
+        public double Interest { get; set; }
 
         public int Years { get; set; }
     }
